fix(updateMethod): always initialize updateMethod
When mocha runs a suite of tests, it runs them all as a single file it
seems. So when updateMethod wasn't being properly initialized, then it
would retain the setting from a previous run (in this case the update
method tests themselves). This lead to inconsistent failures when
running the entire test suite