feat(isRestify): allow user to specify whether restify is used
Express and restify differ solely in their support for the `delete`
method on the server instance (for the sake of thinky-rest that is).
Previously the check to determine whether restify was being used
was to compare the `name` property to `restify`, however this is
allowed to change between instances. This patch adds a heuristic to
determine whether restify is being used, as well as letting the
user explicitly pass in the value for `isRestify` during the
initialization of this module.